+++ config/console/params.php
@@ -6,4 +6,9 @@ return [
     'yiisoft/yii-console' => [
         'commands' => require __DIR__ . '/commands.php',
     ],
+
+    'yiisoft/db-migration' => [
+        'newMigrationNamespace' => 'App\\Migrations',
+        'sourceNamespaces' => ['App\\Migrations'],
+    ],
 ];